ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.28
langchain==0.1.0
langchain-openai==0.0.2
# TODO: Configure LLM provider - add anthropic if using Claude
# anthropic==0.8.1
openai==1.6.1
pymilvus==2.3.4
grpcio==1.60.0
grpcio-tools==1.60.0
pydantic==2.5.0
numpy==1.26.2
# TODO: Choose embedding model
sentence-transformers==2.2.2  # For local BGE models
# torch==2.1.1  # Required for local models
# transformers==4.36.0  # Required for local models
redis==5.0.1
prometheus-client==0.19.0
pyyaml==6.0.1
